requirements.txt
selenium==4.1.5
robotframework==5.0.1
robotframework-seleniumlibrary==6.0.0